Version 1.12
Stack and Max buttons display a confirmation regardless of purchase size
to alleviate accidental clicks.
The BuyEmAll frame will not hide until you accept the confirmation or if
there is no confirmation.
Fixed a bug where you could attempt to buy an item a vendor no longer
has.
German translation courtesy of JokerGermany.
Changed the text on Max and Stack tooltips.
Corrected a few typos in this readme.