Version 2.1
Slightly rearranged the buttons to have a bit more space in between.
Clicking on the Max and Stack buttons will now change the amount in the
box instead of immediately bringing up the confirmation window.
Stack button will now be disabled for items whose stack size is the same
as their purchase unit (ammunition, for instance).