Version 3.4.0
Formatted parts for easier readablity, for me at least.
Rewrote the alternate currency handler to be more adaptive.
Thanks to a recent (7.2.5) change, removed my clunky alternate currency
ID function.
Added an auto rounding function to work better with buying a vendor
stack for one item. (e.g., Blood of Sargeras)
A note on the auto rounding, the number you see in the frame will be
rounded to the nearest multiple of the purchasable item when bought.